requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.28
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.9.0
